+++ package.json
@@ -21,7 +21,9 @@
     "@react-navigation/native": "^7.1.17",
     "@react-navigation/native-stack": "^7.3.25",
     "expo": "~53.0.22",
+    "expo-application": "~6.1.5",
     "expo-dev-client": "~5.2.4",
+    "expo-device": "~7.1.4",
     "expo-document-picker": "~13.1.6",
     "expo-font": "^13.3.2",
     "expo-image": "~2.4.0",
@@ -34,8 +36,7 @@
     "react-native-safe-area-context": "5.4.0",
     "react-native-screens": "~4.11.1",
     "react-native-vector-icons": "^10.3.0",
-    "expo-application": "~6.1.5",
-    "expo-device": "~7.1.4"
+    "expo-clipboard": "~7.1.5"
   },
   "devDependencies": {
     "@babel/core": "^7.25.2",
