+++ src/components/index.ts
@@ -1,3 +1,1 @@
-export * from "./analytics";
-export * from "./overview";
 export * from "./ui";
