+++ src/components/sections/analytics/SecurityAnalytics.tsx
@@ -1,8 +1,8 @@
 import { StyleSheet, View } from "react-native";
 import MaterialIcons from "react-native-vector-icons/MaterialIcons";
 import type { EventCount } from "@/modules/dpkg-module";
-import { TText, useTheme } from "../../theme";
-import { TableRow, TableRowGroup } from "../ui";
+import { TableRow, TableRowGroup } from "@/src/components/ui";
+import { TText, useTheme } from "@/src/theme";
 
 export function SecurityAnalytics({ analytics }: { analytics: EventCount }) {
   const { theme } = useTheme();
@@ -43,22 +43,11 @@ export function SecurityAnalytics({ analytics }: { analytics: EventCount }) {
       {securityItems.map((item) => (
         <TableRow key={item.label}>
           <View style={styles.tableRowContent}>
-            <View
-              style={[
-                styles.iconCircle,
-                {
-                  backgroundColor: item.isAlert
-                    ? `${theme.error}20`
-                    : `${theme.primary}20`,
-                },
-              ]}
-            >
-              <MaterialIcons
-                name={item.icon}
-                size={24}
-                color={item.isAlert ? theme.error : theme.primary}
-              />
-            </View>
+            <MaterialIcons
+              name={item.icon}
+              size={24}
+              color={item.isAlert ? theme.error : theme.primary}
+            />
 
             <View style={styles.textContainer}>
               <TText
@@ -96,12 +85,5 @@ const styles = StyleSheet.create({
     justifyContent: "center",
     flex: 1,
     marginLeft: 16,
-  },
-  iconCircle: {
-    width: 40,
-    height: 40,
-    borderRadius: 20,
-    justifyContent: "center",
-    alignItems: "center",
   },
 });
