+++ src/components/sections/index.tsx
@@ -1,0 +1,3 @@
+export * from "./analytics";
+export * from "./overview";
+export * from "./settings";
