+++ src/components/sections/overview/FavChannelsList.tsx
@@ -1,8 +1,8 @@
 import { StyleSheet, View } from "react-native";
 import MaterialIcons from "react-native-vector-icons/MaterialIcons";
 import type { ExtractedData } from "@/modules/dpkg-module";
-import { TText, useTheme } from "../../theme";
-import { TableRow, TableRowGroup } from "../ui";
+import { TableRow, TableRowGroup } from "@/src/components/ui";
+import { TText, useTheme } from "@/src/theme";
 
 export function FavChannelsList({ data }: { data: ExtractedData }) {
   const { theme } = useTheme();
