+++ src/components/sections/settings/index.tsx
@@ -1,0 +1,4 @@
+export * from "./AppearanceSection";
+export * from "./AppInfo";
+export * from "./CreditsSection";
+export * from "./DeviceInfo";
