+++ src/components/ui/index.ts
@@ -1,7 +1,7 @@
 export { ButtonText, default as Button } from "./button";
 export { default as Checkbox } from "./checkbox";
 export { default as Header } from "./header";
+export { default as Toggle } from "./switch";
 export { TableRow, TableRowGroup } from "./TableRow";
 export { default as Toast } from "./toast";
 export { ToastProvider, useToast } from "./toast/provider";
-export { default as Toggle } from "./toggle";
